Extract domain document boundary checks
